+++ tests/test-dirs/locate/context-detection/cd-mod_constr.t/run.t
@@ -3,9 +3,9 @@
   {
     "class": "return",
     "value": {
-      "file": "lib/ocaml/string.ml",
+      "file": "lib/ocaml/stdlib.ml",
       "pos": {
-        "line": 1,
+        "line": 722,
         "col": 0
       }
     },